+++ src/Individual.java
@@ -1,5 +1,5 @@
 public class Individual extends Departamentos{
-    public Individual(int id, double areaMetrosCuadrados,double valorCompra,String direccion,int numHabitaciones,int numBanos, float administracion){
+    public Individual(int id, double areaMetrosCuadrados, double valorCompra, String direccion, int numHabitaciones, int numBanos, double administracion){
         super(id, areaMetrosCuadrados, valorCompra, direccion, numHabitaciones, numBanos, administracion);
     }
     public void mostrarInfo(){
